+++ CandyCoat/Data/VipDurationType.cs
@@ -1,0 +1,3 @@
+namespace CandyCoat.Data;
+
+public enum VipDurationType { Monthly, OneTime, Permanent }